Version 3.27.6 (based on bleeding_edge revision r21348)
Performance and stability improvements on all platforms.

git-svn-id: https://v8.googlecode.com/svn/trunk@21349 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/platform-win32.cc
@@ -1191,7 +1191,7 @@ void OS::SignalCodeMovingGC() { }
 #endif  // __MINGW32__
 
 
-uint64_t OS::CpuFeaturesImpliedByPlatform() {
+unsigned OS::CpuFeaturesImpliedByPlatform() {
   return 0;  // Windows runs on anything.
 }
